validaciones con zod a los endpoints agregadas

+++ src/controllers/usersControllers.js
@@ -1,16 +1,32 @@
 const { User } = require('../models/usersModel')
 const bcrypt = require('bcrypt')
+const { registerSchema } = require('../common/users/user_schema');
+
 
 const userGet = async (req, res) => {
-  const buscar = await User.find({})
-  res.status(200).json(buscar)
+
+  const id = req.params.id;
+
+  if(!id) return res.status(400).json({ message: 'ID requerido' })
+
+  const user = await User.find(id);
+
+  if(!user) return res.status(404).json({ message: 'Usuario no encontrado' });
+  res.status(200).json(user);
 }
 
-const userPost = async (req, res) => {
-  const { name, age, email, password } = req.body
+const userPost = async (req, res, next) => {
+  const { name, age, email, password } = req.body;
+
+  const check = registerSchema.safeParse({ name, age, email, password });
+  if(!check.success) return res.status(400).json({ message: check.error.issues[0].message })
+
+  const user = await User.findOne({ email });
+  if(user) return res.status(409).json({ message: 'Correo ya en uso' })
+
 
   const newUser = new User({
-    //nuevo usuario con schema
+    // nuevo usuario con schema
     name,
     age,
     email,
@@ -26,6 +42,12 @@ const userPut = async (req, res) => {
   const id = req.params.id
   const { name, email, password } = req.body
 
+  const check = registerSchema.safeParse({ name, email, password });
+  if(!check.success) return res.status(400).json({ message: check.error.issues[0].message })
+
+  const user = await User.findById(id);
+  if(!user) return res.status(404).json({ message: 'Usuario no encontrado' })
+
   const userPut = {
     name,
     email,
@@ -37,6 +59,9 @@ const userPut = async (req, res) => {
 
 const userDelete = async (req, res) => {
   const { id } = req.params
+
+  const user = await User.findById(id);
+  if(!user) return res.status(404).json({ message: 'Usuario no encontrado' })
 
   const eliminar = await User.deleteOne({ _id: id })
   res.status(200).json(eliminar)
